ship features part
