+++ apps/docs/app/(home)/_components/features.tsx
@@ -6,6 +6,7 @@ import {
   CardContent,
   CodeBlock,
 } from "@rehooks/ui/components";
+
 import { cards, type Card as CardType } from "@rehooks/utils";
 
 export function Features() {
@@ -14,9 +15,9 @@ export function Features() {
       <h2 className="text-muted-foreground/50 select-none text-xl font-medium">
         FEATURES
       </h2>
-      <h3 className="text-foreground mt-1.5 max-w-lg text-wrap text-center text-xl">
-        Rehooks is created to streamline the process of creating custom hooks,
-        and to make it faster to ship.
+      <h3 className="text-foreground mt-1.5 max-w-lg text-wrap text-center text-xl italic">
+        I've created Rehooks to streamline the process of creating custom hooks,
+        and get rid of heavy packages.
       </h3>
       <Cards />
     </>
