+++ packages/ui/package.json
@@ -32,6 +32,7 @@
     "@types/react": "^18.2.61",
     "autoprefixer": "^10.4.18",
     "class-variance-authority": "^0.7.0",
+    "framer-motion": "^11.15.0",
     "lucide-react": "^0.453.0",
     "postcss": "^8.4.35",
     "sugar-high": "^0.7.5",
