+++ packages/ui/src/components/card.tsx
@@ -65,7 +65,7 @@ const CardContent = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <div
     ref={ref}
-    className={cn("rounded-md p-4 font-mono text-sm", className)}
+    className={cn("relative rounded-md p-4 font-mono text-sm", className)}
     {...props}
   />
 ));
@@ -100,7 +100,6 @@ const CodeBlock = ({
             "--sh-sign": sign,
             "--sh-string": string,
             "--sh-keyword": keyword,
-            "--sh-comment": comment,
             "--sh-identifier": identifier,
           } as React.CSSProperties
         }
