--- packages/utils/src/constants/cards.ts
@@ -1,30 +1,0 @@
-import type { Card } from "../types";
-
-export const cards: Card[] = [
-  {
-    label: "Rehooks",
-    title: "TypeScript",
-    content:
-      "Written in TypeScript, Rehooks offers comprehensive type safety and autocompletion.",
-  },
-  {
-    label: "Rehooks",
-    title: "Performant",
-    content:
-      "All hooks are meticulously crafted with SOLID principles, ensuring efficient, optimized, and robust performance.",
-  },
-  {
-    label: "Rehooks",
-    title: "Performant",
-    content:
-      "Rehooks is designed to be performant, with optimized code and minimal overhead.",
-  },
-  {
-    label: "Rehooks",
-    title: "Open Source",
-    code: `const rehooks = {
-  openSource: true,
-  license: "MIT"
-}`,
-  },
-];
