+++ packages/utils/src/constants/cards.tsx
@@ -1,0 +1,31 @@
+import type { Card } from "../types";
+
+export const cards: Card[] = [
+  {
+    label: "Rehooks",
+    title: "TypeScript Support",
+    content:
+      "Rehooks is written in TypeScript, ensuring type safety and maintainability throughout the entire codebase.",
+  },
+  {
+    label: "Rehooks",
+    title: "Variety of Hooks",
+    content:
+      "Rehooks provides a wide range of hooks for various use cases, eg. useThrottle, useDebounceCallback, useFocus...",
+  },
+  {
+    label: "Rehooks",
+    title: "Performant & Reusability",
+    content:
+      "All hooks are meticulously crafted with SOLID principles, ensuring efficient, optimized, and robust performance.",
+  },
+  {
+    label: "Rehooks",
+    title: "Open-Source Codebase",
+    code: `const rehooks = {
+  openSource: true,
+  license: "MIT",
+  issues: "open"
+}`,
+  },
+];
